CoverageControllers

Documentation for CoverageControllers.

CoverageControllers.Grid

CoverageControllers.Grid

CoverageControllers.Grid

Base.axes

Base.axes

CoverageControllers.PolygonBoundary

CoverageControllers.center

CoverageControllers.dct_trajectory

CoverageControllers.fill!

CoverageControllers.lengths

CoverageControllers.trajectory_map

CoverageControllers.Grid — Type

Grid{D, T, F}

Represents a D-dimensional grid, with origin o, spacing along each axis of dx and N grid points along each axis. The dct_plan is computed since it is used so frequently

CoverageControllers.Grid — Method

Grid(o, dx, N)

construct a grid with with origin o spacing dx and N elements along each axis.

CoverageControllers.Grid — Method

Grid(origin, dx, lengths)

construct a grid with origin o spacing dx and lengths lengths along each axis.

Base.axes — Method

axes(grid, dim)

returns the axes for the grid at the specified dim

Base.axes — Method

axes(grid)

returns an ntuple of the axes along each dimension of the grid

CoverageControllers.PolygonBoundary — Method

PolygonBoundary(points) returns a Vector{LineBounary} representing a polygonal boundary

CoverageControllers.center — Method

center(grid)

returns the center of the grid

CoverageControllers.dct_trajectory — Method

dct_trajectory(grid, traj)

converts a list of visited points to a spatial distribution, and then returns the dct of the spatial distribution

CoverageControllers.fill! — Method

fill!(func, grid, data)

fills the data matrix by applying func to each pos. func should be a function that takes in a position (as an n-tuple) and return a scalar.

CoverageControllers.lengths — Method

lengths(grid)

returns the length of each axis of the grid

CoverageControllers.trajectory_map — Method

trajectory_map(grid, traj)

convert a list of visited points to a spatial distribution